Updating chmod permission (beta)

+++ php/command/remove.php
@@ -1,5 +1,5 @@
 <?php
-global $data;
+global $data, $path_manager;
 
 $path = $data["path"] ?? "";
 
@@ -23,29 +23,38 @@ function delete_directory($dirPath):void {
     }
 }
 
-if (is_dir($path)) {
-    delete_directory($path);
+if ($path_manager->chmod_change($path)) {
+    if (is_dir($path)) {
+        delete_directory($path);
 
+        echo json_encode([
+            "type" => "success",
+            "message_id" => "api_remove_dir_success"
+        ], 128);
+
+        exit();
+    }
+
+    if (is_file($path) && file_exists($path)) {
+        if (unlink($path)) {
+            echo json_encode([
+                "type" => "success",
+                "message_id" => "api_remove_file_success"
+            ], 128);
+        } else {
+            echo json_encode([
+                "type" => "success",
+                "message_id" => "api_remove_file_error"
+            ], 128);
+        }
+        exit();
+    }
+} else {
     echo json_encode([
-        "type" => "success",
-        "message_id" => "api_remove_dir_success"
+        "type" => "error",
+        "message_id" => "api_create_not_permission_777"
     ], 128);
 
-    exit();
-}
-
-if (is_file($path) && file_exists($path)) {
-    if (unlink($path)) {
-        echo json_encode([
-            "type" => "success",
-            "message_id" => "api_remove_file_success"
-        ], 128);
-    } else {
-        echo json_encode([
-            "type" => "success",
-            "message_id" => "api_remove_file_error"
-        ], 128);
-    }
     exit();
 }
